morebits: remove use of "async: false" actions
There is no need to use "async: false" in both the places where this was being used till now. The latter was added by me in #682; don't know what I was smoking when I wrote that.

"async: false" has been deprecated in jQuery since version 1.8. Its usage is not compatible with the use of $.Deferred() in $.ajax(), and thus do not work with the new promise-based Morebits.wiki.api class. Synchronous XMLHttpRequests are anyway bad on performance and should not be used anywhere.