morebits: add Morebits.quickForm.getInputData
New function for extracting all filled-in data from a form.

A small change is made to Morebits.quickForm as well, to help us tell apart the checkboxes
whose value is a boolean from the ones whose value is to be list of checked items.

Optional reading, for those who want to know why the quickForm modification is necessary:
"Wrap in multiple issues" (tag module) is a checkbox for which the value should be `true` or `false`.
On the other hand, if we have a list of checkboxes (each having the same name parameter) from
which the user checks one or more, we want the result to be an array of the values of the
checked elements.
It does not suffice to just check the number of checkboxes (1 or more) and decide whether
to use boolean value (in case of 1 checkbox) or array value (in case of multiple), because
there are cases where the number of checkboxes is not known in advance (like in the batch
modules) - here if there is only one wikilink on the page, there will be only one checkbox,
but we nevertheless want an array here, not a boolean.