talkback: Remove parsing of html
Template:No_talkback creates a link to `http://userjs.invalid/noTalkback?reason=`; that gets used by the talkback module to provide a custom reason to the user, but by using `.append`, it allows someone to provide and run code via `<script>`.  Changing to `.text` does away with that.  See T241950.

In the future, there are other options here, such as querying `templates` (removes the ability for editors to provide a custom `reason`), querying `revisions` (regex to match page the content), or simple removing the ability to provide custom text.