talkback: Remove custom optout message
First added in 2013 (99893a8).  Relies upon querying `extlinks`, but 6.5 years later, only one three enWiki users (out of 26 or so) use the custom message, only one of which is an active editor (@JJMC89, and doesn't mind this being done away with).  The alternative is querying `revisions` and matching a regex, but that's no fun.

With this we can blank Template:No_talkback.

OR: we make it add a category and just use the category query.  Probably better.