Block: Tweak reblock message when user is partially blocked
This merely notes whether a user is partially blocked or sitewide blocked in the red message to sysops.  Also notes in a few minor places that the module is for sitewide blocks (#802).

A query to `list: 'blocks'` without any `bkprop`s defined (as we have here) apparently returns `partial: ''` if the user is partially blocked, and nothing if they're not, so relies on that.