arv: replace use of moment with Morebits.date
No loss in functionality.

As an enhancement, the relative times are now shown in the user time zone (from the user's setting at Special:Preferences) rather than in the system's time zone. This is in line with how the mediawiki core and extensions work. The timestamps which are saved to the wiki remain in UTC time zone, of course.