Remove Twinkle.userAuthorized
It's not really serving much of a purpose: used only in `twinkle.js` and `twinklefluff.js`.  Consolidated in 6dc2a75 (2013) as part of #151, where the requirement was added to the gadget definition.  It used to be checked in each module (ad5af2f, 2011) but is now left just in `fluff`; there's some sense to that, though, as `fluff` is the only one that doesn't add a menu button (see also #792).

The reality, though, is that `Twinkle.userAuthorized = false` is not much of a "security" check, all the moreso because it came in `Twinkle.load`.  Moving the check higher up, before the Twinkle object is declared, makes things somewhat more tricky.  We could be slightly trickier, forgoing `Morebits` and doing:

`!(mw.config.get('wgUserGroups').indexOf('autoconfirmed')!==-1 || mw.config.get('wgUserGroups').indexOf('confirmed')!==-1)`

but the reality is that anyone trying to get past it will have no difficult doing that as well; indeed, anyone seriously attempting this far would probably just copy/write their own.

I've removed the check in `fluff` but not replaced it with anything; thus it's possible for someone to somewhat easily get the links, but, as above not particularly difficult or protective.

I've also removed the message in `Twinkle.addPortlet` as I couldn't trigger that message.