xfd: Specify CfD action in cfd-notify
Be specific with what action (deletion, splitting, etc.) the user is performing for a given CfD nomination.  I recently [changed](https://en.wikipedia.org/wiki/Template_talk:Cfd-notify#Put_action_in_section_title) {{cfd-notify}} to make use of `action` in the title as well as the body, so this should be doubly helpful.

This changes the edit summary on Today's log page from "Adding delete nomination" to "Adding deletion nomination" and so on, which I think is for the better anyway.