warn: Fix bug with kitchensink edit summaries
Reported at WT:TW (https://en.wikipedia.org/w/index.php?title=Wikipedia_talk:Twinkle&oldid=939908648#Apparent_(rather_minor)_Twinkle_bug).  Basically, when #773 added the kitchensink menu, I didn't account for the fact that the edit summary test was relying on the `main_group` to have a number in it, which fails with kitchensink.  For kitchensink, then, we rely on the `sub_group` via `substr`, as with custom; similarly, it's over-complicated because of 4im levels.  Also worth noting that `uw-3rr` is a non-level1-4im with a digit in it.