morebits: lookupCreation won't query content and section unless lookupNonRedirectCreator is provided
PR #710 added the ability for `fnLookupCreationSuccess` to find the creator of the first non-redirect revision via `fnLookupNonRedirectCreator` (see also #805) by adding `content` to the `rvprop`s list as well as `rvsection=0`.  That doesn't work for non-wikitext content models, who will returned a `rvnosuchsection` error in such cases.  Removing `content` would solve the error, as would `rvsection=0`, but the actual redirect check (`/^\s*#redirect/i`) doesn't make sense in those scenarios (e.g. Scribunto, JS, CSS), so there's no really point in even checking.

Instead, only attempt to use `content` and `rvsection` if `lookupNonRedirectCreator` is set.  Currently, only AfD and PROD do so, which aligns with how it should be limited.

We could probably just use `wgPageContentModel`, but in theory the ability to query pages other than the current one, even if the two have different content models, seems useful.  In practice, only `twinklexfd` attempts sch a thing, only for TfD merge nominations (`involvedpages`), and assumes that the target is in the same namespace as the current page, so Scribunto will match with Scribunto.