xfd: reenable setPageSection for mfd and cfd
Disabled in c97229c due to bugs in the API returning edit conflicts; temporally, it seems related to #19, and I think 45ce088 may have been trying to fix the same thing.  I'm not sure what the error was but that was nine years ago, and testing seems fine, so I'm hoping this is okay, sort of a test for the rest.