xfd: Avoid client clock for soft redirects
In doing #862, I undid the intent of #714 to avoid the client clock as much as possible for the specific case of soft redirects being nominated at RfD.  It never worked - thus #861 - but the idea is sound.  This has softredirects go through the same `curtimestamp` query and "typical" redirects, but simply skips processing of target or section.  It's not an entirely free check anymore, but it should be incredibly infrequent and fairly quick.