speedy: Log nomination even if user notification prevented
Since #578, we've been trying to correctly note in the CSD log whether the user notification went successfully.  That was carried over into #563, but the changes in #578 meant that if the user attempted to notify the creator (e.g. `params.usertalk`) but then discovered (in `noteToCreator`) that there was a good reason not to, no CSD log entry was made; it accounted for whether you wanted to notify-then-log or just log, but didn't account for notify-but-notify-not-done-then-log, essentially ignoring the checks from #436.  Specific cases, like R2 (reported at WT:TW http://en.wikipedia.org/wiki/Special:Permalink/942392393#CSD_R2_instances_not_being_logged), were especially prone to being missed in the log.

This uses those `initialContrib = null` checks to recheck whether a notification should go out, and if not, log if appropriate.