speedy: Fix long-standing bug in G6 copypaste notice
I only noticed this after looking at things after #849, but we've been giving inaccurate notices for G6's copypaste for ages.  {{db-copypaste-notice}} is just {{uw-c&pmove}} at a different title so Twinkle can handle providing the notice, but that notice uses a different format than the rest of the (standardized) csd notices.  In particular, parameter `1` is *not* the page name, but rather the name of the *original* page.  I *think* it's likely fine to change {{db-copypaste-notice}}, but I'm not sure if anyone else uses it and for now this is simpler.