morebits: don't handle notoken error
notoken error code has been removed from MediaWiki. Write queries
without token now return the standard missingparams error.

It is anyway not necessary to handle this error since we do specify
the token, so this just wouldn't ever occur.

In case of expired tokens, the `badtoken` error continues to be
returned, which we do handle.