xfd: Add XfD log, parallel to CSD and PROD
There's not the same reason to log XfD nominations as there is for CSDs and PRODs (see #303 where it was dismissed), but this is useful/interesting regardless; a few folks have asked for it or thought it would be useful.  It is also a lot more complicated.

The basics:
- Adds config options (logXfdNominations, xfdLogPageName, and noLogOnXfdNomination)
- Users can set preference not to log a given venue
- Various venue-specific log formatting (`toTLACase`, etc.)
- Like PROD/CSD, checks whether user was notified (although see below).  CfDS and RM have no user notification.
- Includes a link to the page log for files
- Stores common parameters and makes use of `.extend`
- Uses new `Morebits.userspaceLogger` from #878

Nitty-gritty:
There are real complications around several venues, in particular confirmations on whether something actually happened and what information is available.  Most are dealt with, some aren't.  At a very basic example, we're just simply not going to check the success of each and every deletion sorting listing, even if we wanted to add them to the XfD log, which gives some leeway for being lax on other parameters.  To wit:

- RfD logs the notification of just the initial contributer, not the target page (although it does log the redirect's target)
- MfDs of userspace pages created by a different user will log both names, but only checks if the notification to the creator was successful
- TfD merges (TfM) only log the creator of the first page, and can't currently note whether the second template's creator was notified or not (it *does* log the name of that second page, though)
- Stub template nomination (#897) three-letter-acronyms (SfD and SfR) aren't commonly used or understood.