morebits: Remove twinkle-specific CSS that is 100% redundant
See #778.  The stuff in the portlets section was all added in 511bd8f and is just for the Modern skin, but it is all already present in the main Modern CSS: https://gerrit.wikimedia.org/r/plugins/gitiles/mediawiki/skins/Modern/+/master/resources/main.css.  Don't think it was even needed at the time, but regardless, does nothing and can be removed.