lint: Enforce quote-props consistent-as-needed
Not done in #500 given the cosmetic nature and fairly mixed status of the project, but I've run into it a few times lately and been mildly annoyed.

Adding `"keywords": true` seems reasonable, but it's not that common and I don't think in the end we really care.  Likewise with `numbers`.

See https://eslint.org/docs/rules/quote-props