xfd: Better fixes for namespaces in cfd/tfd/cfds previews and form replacements
Better and more complete way of doing 133d227 (#857), for both CfD and TfD, as well as some tidying for CfD/S.  Only replaces the current namepsace, which is basically what was being done anyway, but this way it's obvious when we're pigeonholing a page into an incorrect namespace.  For CfD/S, the behavior is basically inverted: instead of removing the namespace then adding it for just the discussion page, this removes it just for the tagging.