morebits: Change patrol to query page info and use revid if rcid link not available
Should fix a lot of missed patrollings.  First added in cae9a736 (Apr 2011), this was before the `revid` option was available (added in MW 1.22, circa 2013).  The old `rcid` method would fail if the link didn't exist, such as when not done from the page itself where the "Mark as patrolled" link providing the `rcid` was available (history or edit actions), or if the page was too old to be in the recent changes table.

Format is a little non-standard, mainly because we don't really need to query unless necessary and because we're intentionally ignoring any errors.

Also improves the "poor man's normalisation" with `mw.Title`.