fluff: Add ability to enable rollback links on Recentchanges(linked)
Adds new `recent` option, off by default.  Most of the complexity in the selectors comes from having to handle two preferences: using the new or old (no js) RC system, and the grouped or ungrouped preference.  Uses the `wikipage.content` hook to (re)generate links when new diffs are loaded.

Also expands `openTalkPageOnAutoRevert` to apply at RC/RCL, just like with contribs.  The text for that has been tweaked as well, as it was a misleading description around behavior for contributions pages.  The language stemmed from 0746c05c in 2011, but it wasn't an accurate description of how this preference is helpful.  It *does* define whether the user talk page is opened, but it doesn't help do multiple things on the contributions page, because a single revert will still redirect you away from the contributions page (and now recent changes).  That is itself perhaps undesirable, but means this was misleading.