fluff: Restructure twinklerevert links to use Twinkle.fluff.revert, remove auto function
As far as I can tell, all the desired information is already available, thanks to some classes, so there should be no need to unnecessarily load a page.  The revert query itself should be able to handle the case where someone else has edited the page, so we don't really need the `auto` behavior now that we have the information ready.  The other loss is handling of pages that fail `wgIsProbablyEditable`, but this is manageable by making use of `intestactions=edit` (see also #632 and #658).  By not relying on auto-processing `twinklerevert` urls, we can prevent malicious providing of a link with `&twinklerevert=vand` to entice someone to revert something.  Low-profile, and I've never heard of it, but possible.

Given the above, `autorevert` has been removed in favor of a `skipTalk` item modified on contribs and recent changes to more appropriately represent what `openTalkPageOnAutoRevert` does.  That config is unfortunately named, but isn't visible so shouldn't be an issue beyond maintenance.