fluff: Restructure to use one function for rollback links
Not strictly necessary but it's clearer to not separate functions because of largely artificial differences.  Converts one function with 3 parameters and another with 2 to one with 2-4 parameters.

NB: This changes the new links on history pages (#942) from three to two rollback links, as all inline links are now done as pairs rather than triplets (the latter now only showing up on diffs).  Also renames the red vandalism link in the triplet/on diffs from `rollback (VANDAL)` to `vandalism`, matching behavior elsewhere and taking up less space overall.