Decentralise module initialisations
Initialise the modules from the module files' themselves rather than centrally from twinkle.js. Doing it
this way makes testing of twinkle.js and of custom Twinkle installations (like on other wikis) less
cumbersome as one no longer needs to comment out or modify the initialisation lines in twinkle.js (as mentioned [here](https://en.wikipedia.org/wiki/Wikipedia_talk:Twinkle/Archive_42#How_to_use_as_a_user_script?) and [here](https://github.com/azatoth/twinkle/pull/462#issuecomment-462495588)). This also fixes the issue https://github.com/azatoth/twinkle/pull/944#issuecomment-627359133 allowing for automated setup for a test file for changes that affect twinkle.js.

To do this, we re-purpose Twinkle.addInitCallback() function so that it takes the name of the function as
well, so that we can check if the module was disabled. For twinkleconfig and any "custom modules", the
disabled check is unnecessary so no name needs to be provided.

The order in which the menu buttons appear is still within our control through MediaWiki:Gadgets-definition - the files are executed in the order in which they appear in the gadget definition line.