Remove useragent 2.0 specification
First added in c236127 but the 2.0 specification hasn't been updated or considered in nearly a decade and its not something we're using.  These are used for logging purposes, all that really matters is that there's a unique and identifiable string, for which the tool and project suffice.

Also move to `wgWikiID` rather than `wgDBname` per https://phabricator.wikimedia.org/rMWd45baf7f0734a2506c77925064f42bc13e1bff5e