block: Fix bug showing erroneous input box on a reblock form initialization
The `area` box would show on initialization if the user was already blocked, when it should only show iff the user is receiving a partial block template but *not* a block.  Changing the template or preset would kick the checks into action, but they weren't being processed on a reblock.  Original sin is from the initial adding of partial blocks in #813/61b5269.