Apply twinkle change tag to all actions
Fairly rote.  Added `summaryAd` to twinkleprotect's `stabilize`, since that was missing it (T247721); didn't do the same for `triage` since while `pagetriageaction` has a `note` parameter that functions like an edit summary we could use, it felt like overkill to add something to morebits just so we can modify `ctx.note` for these largely pointless log entries.  Ideally, https://phabricator.wikimedia.org/T252980 will be taken care of first.  There are a few instances where `setChangeTags` is used earlier than might otherwise be expected, so as to apply to downstream options; a number of these are the `triage` calls, which are thus, as noted, nonfunctional.