morebits: Only triage or patrol if page is available for doing so
Fallout from #917, in particular with the `triage` action.

`triage`: When AfD-ing an article that was already reviewed, Twinkle would throw a scary bold error saying that the article wasn't in the queue.  Which, of course, is true.  We could check the API each time to see if the apge is uncurated (e.g. introduce `triageProcessListApi` as an intermediary), but there's an order-of-operations annoyance there, since `pagetriagelist` and `pagetriageaction` both require `pageid`, so we would *have* to query for that first and wait.  Instead, this just ignores the specific error (`bad-pagetriage-page`); it will still output a note in the console, but should be faster/smoother on the user end.  See discussion in https://github.com/azatoth/twinkle/pull/930.

`patrol`: Doesn't throw up an error, but just notes the action as successfully done.  We can query the `recentchanges` list in the same query wherein we get a token/etc., although doing so removes the option for a query-less patrol action when we're on the page but the rcid link isn't available, e.g. history.