warn: Don't follow cross-namespace redirects out of usertalk space
Option added in #915.  This one requires some thinking since we have to deal with autolevel's expectation that a valid page object will be loaded.  We really just need to catch it once the menu has loaded - or tried to - and insert a strong note there.  It's somewhat equivocating, as presumably there could be other issues.  Also closes any preview that was previously opened.