morebits: Remove tipsy dependency, use jQuery UI for tooltips
This harkens back to the homegrown tooltips that were replaced by tipsy (a14f46f, 2011), mainly because we've still been defining the CSS!  Some tweaks were made: with a decade of black, I figured blue would be too bright, so mediumblue was used.  The classes are changed, but there should only be minor differences in appearance.

https://api.jqueryui.com/tooltip/