prod: Fix a number of bugs around endorsing
Most prominently, this moves the check for `{{old prod}}` to after the endorsement check, which allows proper endorsing (on the fritz since #735).  This means that if there's a prod already present, this will allow endorsing a (regular) PROD even if the talk page has been tagged (as should be the case).  This will also now *not* add another old prod template to the talk page when BLPPRODing.

Additional fixes:
- Correctly note prior tag type when endorsing
- Move article age check into new prod check, where it makes the most sense.