fluff: Fail nicely on and don't reveal revdel'd usernames
Sometimes we use or reference usernames in fluff, whether the user being reverted or occasionally the user of the revision to which we are reverting.  This would cause issues if the username was hidden; more alarmingly, if the twinkler was a sysop, the username could be inadvertantly revealed (same for oversight although we're pretty careful about clean up).  This basically checks the `userhidden` property on the selected revisions, then uses that to determine whether to display the user name or a placeholder text (present in the new (as of this commit) `Twinkle.fluff.hiddenName`, set to `an unknown user`).  There is some slightly extra infrastructure around `params.user`, given that the selectors mean we're passing a falsey value for the username, but I think it's clearer this way, and opens the door for future works around cases like diffs viewed with `&unhide=1`.

The main annoyance is on Special:Contributions, where someone with the appropriate group would be able to see revdel'd edits.  The only difference for those edits is a `strong` element with no span or class selector available, so in order to selectively remove these, we'd have to do something like searching for all `strong` elements, then check the content of each one, as there are other bolded elements and nearby items (rollback, tags) are not reliably present.  We could manually add something to [[MediaWiki:Rev-deleted-user-contribs]], but that's far from ideal, especially given how rare this likely is.  Thus, it's not done, at least until https://phabricator.wikimedia.org/T255903.

I also improved the `formatSummary` function a bit.  It now checks if the `$USER` string was present before attempting to replace it.  It likewise now checks that a username was supplied before constructing a link; if not, it will use `Twinkle.fluff.hiddenName`.  I also bumped the edit summary limit (done for tag in 009c80e/#886, see https://phabricator.wikimedia.org/T188798)