morebits: Add Morebits.wiki.user class for user info and processing
We have `Morebits.wiki.page`, but a user-specific class was missing from the library, meaning common actions like (un)blocking would be left to bespoke queries with `Morebits.wiki.api`.  Even more commonly, finding out information about a given user is often desired and those (equally-bespoke) queries get complicated depending on what's desired.  The ability *not* to notify bots or indef-blocked users is oft-requested, as is the ability to opt-out of Twinkle notifications.  To do that we need to actually know something about the user, for which `Morebits.wiki.page` is insufficient.

Key features of the `Morebits.wiki.user` class include:
- Simple loading of user info, allowing tools to easily access useful information (blocked status, user groups, registration time, etc.)
- Methods to block and unblock a user, as well as the ability to change user groups
- A `notify` method which checks users in order to skip bots or sitewide indef-blocked users, as well as usertalks with given templates or an opt-out link.

Its structure is very similar to `Morebits.wiki.page` but there are a few subtle differences and improvements.  The most notable difference is that here the user object is required to be `load`ed, so the methods, rather than checking if we need to query for protection/token/etc., will instead autoload the user if need be.  The only exception is when `notify`ing and you aren't checking for bot or indef-block status or optout templates/links; there's no need to `load` the user in that case.

Other things to note:
- The `load` success callback is optional, which while unlikely to make a difference writing tools should make single uses or testing easier
- The `notify` method uses `Morebits.wiki.page` (`append` or `newSection`) behind the scenes
- (un)block and user groups methods share success and error callbacks
- Added `Morebits.botUsernameRegex` to standardize a regex to match a "bot" username, set to `/bot\b` (see also #858 and #876)
- Answers/solves #593 (What bots should we avoid notifying?) as "all" (by default)
- Class is supported in `Morebits.batchOperation`